[custom renderer] Implement compose state object renderers
Basically, for a given compose state object, its underlying
value (by invoking "getDebuggerDisplayValue()") determines how
it's rendered in the "Variables" pane. This is to provide
an auto-unboxing experience while debugging, that users can
identify the data by a glance.

E.g.
1) if the underlying value is an integer `1`, the label is
rendered `1`.

2) if the underlying value is a list, then the given object
is rendered by a `List` renderer instead of the original
`Kotlin class` renderer. That is, "size = xx" is the
label, and the `ArrayRenderer` is the children renderer
in this case.

3) if the underlying value is a map, then the given object
is rendered by a `Map` renderer instead of the original
`Kotlin class` renderer. That is, "size = xx" is the
label, and the `ArrayRenderer` is the children renderer
in this case. When expanding, each of the entry is
rendered by the `Map.Entry` renderer.

TODO: Since such custom renderers are tied to the implementations
on the compose side. We need to ensure contracts are not broken,
say, we need to make sure there's `getDebuggerDisplayValue` method
implemented/not removed.

Bug: 193542973
Test: Added
Change-Id: I574838e80d954b2168289812ad61f407dd9e0355