[AUA] Test recommended / forced upgrade a bit better
Add a service for invoking the Upgrade Assistant UI, and for creating an
appropriate processor.

The only benefit to this service is that it is overrideable in tests, so
do that in order to check that the various entry points actually end up
creating a processor and/or the tool window UI, as appropriate.

Bug: 159995302
Test: existing tests adapted and strengthened
Change-Id: Ia359d93715f06fe0ac149a84623b1fd659cbc175