[BA] Add warning for Jetifier usage
Implement new JetifierUsageAnalyzer. It checks the state of
android.useAndroidX and android.enableJetifier flags, as well as if the
previous checkJetifier task results are available, and produce the
corresponding output.

This CL also adds the UI to surface the result of JetifierUsageAnalyzer.
New page is added to warnings. It shows a button to run checkJetifier
task and presents the result of previous run if available. Currently
we keep the result in memory while studio is running, it is planned to
persist it in IJ project data folder in the follow-up CLs.

The result of checkJetifier task is a list of libraries that require
jetifier and a path from each library to the support library
it depends on. In this CL for the simplicity we present it as a list
of reversed paths. We plan to replace it with a reversed folding tree
in a separate CL.

Bug: 177051531
Bug: 194299408
Test: added JetifierAnalyzer test, updated UI tests
Change-Id: I68dfe1d6aaa9a9b8370ca351d39fa36b0db215db

GitOrigin-RevId: 1b4fe286da2b304bf10e20033687befd63718f97