[Dsl] Add VersionConstraint to the ModelEffectDescription
Synthetic ModelEffectDescriptions, for instance to reflect a
user-initiated change to the model, will have `null` for their version
constraint, indicating no constraint.

However, the ModelEffectDescriptions attached to Dsl elements as a
result of parsing should preserve any version constraint associated with
the particular entry in the external to model tables, so that subsequent
code can detect the use of operators or properties that are deprecated
or no longer valid.  Therefore, as well as using the version constraint
to determine whether a particular entry in the tables is valid, have the
ModelEffectDescription contained in those tables also reference the
VersionConstraint.

Bug: 194192720
Test: will be exercised in Upgrade Assistant
Change-Id: I63f2549c04e7b9037abf22a1e0fab75d3c946784