[Dsl] Add version constraints for deprecated operators
This is not the complete set of operators deprecated in 7.x (and
to-be-removed in 8.0), but it does include all those whose semantics are
straightforwardly expressed (i.e. those without needing special-casing
using the OTHER semantics).

These version constraints are not directly testable, as the Dsl model
will not use these properties in any case (so there's no direct
difference in behaviour as a result of these constraints).  They will be
indirectly tested by the Upgrade Assistant processor for rewriting
deprecated operators.

Bug: 194192720
Test: existing
Change-Id: Ia1f12b50e7dcc9cd290582fcc8ad1d0e33188a8a

GitOrigin-RevId: accc44921509fd689eeb08e7394426a392d7e167