[AUA] Support rewrite of deprecated operators
This supports all the operators with straightforwardly-expressed
semantics.  Subsequent CLs will address those operators with slightly
more bespoke semantics, and will also address the issue that the current
implementation will arbitrarily re-order the operators within their Dsl
blocks.

The behaviour of the processor on Groovy
`buildToolsVersion(...)` (leaving it untouched) is because even though
the helper method of that name is removed from the interface, the Dsl
decorator will reintroduce a setter method of that name for the
interface property, so that syntax will remain legal.

Bug: 194192720
Test: added
Change-Id: I2383c3e5ee484ab0eab71606fc0cc1065c71e86b