Dominators for Client-side Heap Analysis
Add the ability to compute the dominator tree of the heap graph and
output it in async-profiler format to be rendered as a flame graph. The
dominator tree is useful in that it exposes effective ownershp relations
between objects, in that if object A dominates B, then if A is removed,
B becomes unreachable. Tree structures are more amenable to
visualization than arbitrary graphs, so hopefully this can provide a
useful overview of the large structures in the heap.

The dominator computation is based on an iterative algorithm by Cooper,
Harvey, and Kennedy, with some additional optimizations. The number of
iterations required for convergence is bounded by the "loop
connectedness" of the graph (the maximum number of back-edges in any
acyclic path), which is difficult to estimate. In practice around 500
iterations is typical. If the iteration cap is hit (currently set at
2000), some (likely very few) objects will have incorrect dominators.

Change-Id: I4c5098617859b49c13d91cc0fb0cdfbebe255311