[AUA] Implement a rewriting PropertiesOperationInfo
The check for whether a particular property needs rewriting is whether
the current syntactic form (as read from the VersionConstraint field in
the ModelEffectDescription on the Dsl element) supports arbitrarily
future AGP versions.  We can't use the `new` property on the Component
Refactoring Processor, because we might wish to do this as an optional
step on an upgrade from e.g. 7.1 to 7.2, for operators which will be
removed in 8.0.

Bug: 194192720
Test: added to non-equality test; functionality exercised in next CL
Change-Id: Ife8c34c0009c1a028b15415bf0a8534e7670437b