[Templates] Don't use dynamic build tools version
Before this change, a new build tools release would trigger Studio to
generate projects with an explicit build tools version, which might
cause things not to work out of the box.

If a user is using Android Studio 4.2 and a newer preview version of
Studio side-by-side, they might end up installing build tools 31.  This
will trigger Android Studio 4.2 to start generating projects with
buildToolsVersion explicitly specified as 31, which is incompatible with
Android Gradle plugin 4.2. See b/186605878.

Going forward, it's better to have the templates just use a known-good
configuration, so they generated projects that work out of the box.

Bug: 186605878
Bug: 190816401
Test: Removed GradleBuildSettingsTest, Manual
Change-Id: I4fb756fbaf7d76bc6230323bd3308b16caa722aa

GitOrigin-RevId: b75b31b63a5b54d46c658a354709d01a8730ad4b