[BA] Move computation of GarbageCollectionAnalyzer result
Move computation of some of GarbageCollectionAnalyzer results from
AnalyzersProxy to the result class itself to be able to test it
separately.
We need analyzer tests to either run the full flow including build
to test things all-together or touch only the particular analyzer
related code. Running things 'in between' like it was before here
require extra unnecessary mocking and become a constant source of pain.

Bug: N/A
Test: test refactored
Change-Id: I7335cf75b3d93d1c78b78fd62f3a1f40d7f61c0a