[Dsl] Resolve value when considering SdkOrPreviewTransform
Since we have access to the element from a ReferenceTo, if it is an
expression we can get its value and use that to determine whether we
should use the Sdk or Preview property.  This does not eliminate all
possible uses of the fallback setter (or, in future, the need to guess)
but many more cases will be correctly handled.

Bug: 184935026
Bug: 194192720
Test: existing test adjusted
Change-Id: I75b5bb1dc0ff6c92d6079bf376f83c7aa5a08f68

GitOrigin-RevId: 5dc8a894884451018c02c1ccf3bd45063b4f70bd