[Dsl] Improve implementation of ADD_AS_LIST semantics
We create a synthetic GradleDslExpressionList to hold the elements of
the list (corresponding to the argument list of the build file language
operator).  That element needs to be associated with the semantics of
the model operator, so that downstream code can reason about it, and
needs to have a REGULAR property type (not DERIVED) so that if a model
associated with that element is used to delete and re-create the
element, the backends can generate the correct syntax.  Individual list
elements retain their DERIVED nature.

Bug: 194192720
Test: added
Change-Id: Id47da45bc47ec895d07402c26bbbdbbd45d4471b

GitOrigin-RevId: 8c5ce4583c1ea382269295081bafa65ed2e550c4