[Dsl] Make AUGMENT_LIST a bit more general
Pass the model effect through to addToParsedElementList.  The model
effect needs a bit of massaging for the backends to do the right
thing, setting it to CREATE_WITH_VALUE to abstract over the different
ways that addToParsedElementList can be invoked.

The Groovy deletion code needs adapting too.  We now need to delete
empty method calls, but "empty" is tricky: closure arguments aren't
counted as part of the argument list, so we might have an empty arglist
but a valid closure argument, and under those circumstances we must not
delete the argument list, otherwise we will violate Groovy Psi internal
consistency requirements.

Adjust the model maps for operators acting on the resConfigs product
flavor property to use AUGMENT_LIST directly, removing their
special-case treatment.  This causes the delete-and-recreate code path
to trigger on structural modifications of resConfig values, which is
fine but necessitates adjustment of some test file expectations.

Bug: 194192720
Test: existing tests adjusted
Change-Id: I67289c595e6368bac36d577eb456fab4c2d48e88