[Dsl] Make implementation of proguardFiles less bespoke
Use AUGMENT_LIST, and the new CLEAR_AND_AUGMENT_LIST, to replace the
by-hand implementation of proguardFile, proguardFiles,
setProguardFiles (and the analogous consumer- methods) with the generic
ones.  Because this ends up triggering rewrites of the property if it is
structurally modified, this changes some test result files into
semantically-equivalent but textually-distinct forms.

The triggered rewrites also unveil a deficiency in the backend
implementations of createDslMethodCall() -- apparently, until now, it
has been impossible to create a method call with more than zero
arguments: any arguments have simply been ignored.  This came up in this
change because a common element of the proguardFiles property is
getDefaultProguardFile("<something>"), represented as a
GradleDslMethodCall.

(Sadly, this does not fix the invalid Dsl emitted if the proguardFile
property is set directly to a reference to a list, which will continue
to use the proguardFiles operator rather than assignment or
setProguardFiles.)

Bug: 72853928
Test: existing tests adjusted
Change-Id: I69f55e98c3207760b5924dd8d35819f7fc79afe1

GitOrigin-RevId: dbf7e6f166897c19b24ecab4ce4e1b4a1f2d31cc