[Dsl] Simplify inheritance of model maps
Adjust the model map collector to take parent maps as arguments, and
fill the newly-constructed (child) model map using their contents as
well as the stream being collected.

In a theoretical change from the current behaviour, put the child's
entries in the map before their parents', so that if necessary child
elements can override the things specified in the parent maps.  This
should never matter, but it should be less surprising if it ever does.

Bug: n/a
Test: existing
Change-Id: I7a26f78deda1538c6d562be63a556e187c959273

GitOrigin-RevId: 427dfe211819de5234a428fa3362b062f0760786