Create Symbol Source Abstraction
With a need to provide symbol directories without a project, this change
creates the SymbolSource interface, an interface used to collect symbol
directories from different sources. The SymbolFilesLocator has been
updated to wrap SymbolSource as a means to provide higher-level
operations.

In a later change we will change the stand-alone profiler to provide a
SymbolSource or SymbolFilesLocator, allowing the code that initializes
the profiler to define how we collect symbols informations.

Bug: 188223481
Change-Id: I0e1cedcb136de33c1082765eb8c574c8832b262b

GitOrigin-RevId: 48aabc6782684894373fc4b6c94e1312269fb9e3