[Dsl] Move sync-related utilities from core.tests
Neither GradleFileModelTestCase nor GradleSyncMessagesStub needs to be
in the principal tests module, and their presence causes a unnecessary
dependency on the (large) intellij.android.core.tests module in test
modules that use them: at least the Upgrade Assistant tests.

Removing that dependency implies adding a `plugin.xml` file for the
Upgrade Assistant test suite in order for the test environment to be
properly initialized and for all the extension points to be correctly
populated.  It also means having to add a dependency to gradle.dsl from
intellij.android.testFramework.

Bug: 195135367
Test: existing
Change-Id: I3ce87290ea21770bf1abe3f23dfe78e4c8c58b0e