[Dsl] Make manifestPlaceholders a MUTABLE_MAP
Since we have no (legacy) implementation of addToParsedElementMap on
String property names, we have to ensure that the default element
created to hold manifestPlaceholders by dereference has a suitable
ModelEffectDescription.

In addition, to allow the existing parser tests to pass with the legacy
assignment syntax to manifestPlaceholders, we must keep the VAR (or
VAR_BUT_DO_NOT_USE_FOR_WRITING_IN_KTS) entry above the new VAL one,
because the parser does not use VersionConstraints for validating
entries, but uses the first sufficiently matching one.  This is broadly
OK since VAR admits a superset of VAL's operations.

Bug: 148657110
Bug: 194192720
Test: added, existing test adjusted
Change-Id: I385ba31b05d3d87e31d890caec4580e5dd246162

GitOrigin-RevId: 6990d63c2694074fa3f80fefcf62ea34f7cdf460