Add Android facet to every Android module...
and remove the redirection for the Android facet's getInstance method.
This change is mainly for module per artifact where we will have the
Android facet present on every module that is dirived from the Gradle
project.

In order to keep snapshots predictable we also need to revert a change
to AndroidRunConfigurations so that the names come from the module names
rather than the holder module.

Bug: N/A
Test: snapshots updated
Change-Id: I2d96875807d25606341f629e70b075fd04293ff1

GitOrigin-RevId: 9bffe54c8bb49501e625233a9f1b68e0c83535ba