Disable bundle signing when opening APK project
We currently show a GenerateSignedApkAction in the build menu, showing
users a wizard where they can select if they want to generate a signed
APK or bundle.

This change modifies the action to say only sign APK if the project
opened is an APK and in that case the wizards jumps directly to the APK
signing step.

Bug: 186124820
Test: GenerateSignedApkActionTest
Test: Manually confirmed the menu and wizars change accordingly.
Change-Id: I077f55909c3072059f5c803b3dde9097b4275c53