[Dsl] Make followElement work on all GradleDslElements
It already did, and (judging by a comment above resolveElement) was
always intended to, but its argument was more restrictively declared.
It returns the provided element if it is not a reference, so there is no
need for pre-checking before calling; it returns null if the argument is
a reference but the injection is invalid in some way.

Bug: none
Test: existing
Change-Id: I9c3ba811241031dc4186b306e805316f88f7a31e

GitOrigin-RevId: 774af3d2b35d20a7cdcf006495bf892a0c612ea0