[Dsl] Make sure to delete a literal's name element
Most of the time, deleting the literal's expression will, through the
process of deleting semantically-empty PsiElements traversing upwards
through the tree, lead to the PsiElement corresponding to the literal's
name being deleted too.  (For example, the operator name of a Groovy
application statement, or the left-hand side and operator of an
assignment statement).

However, in some rare cases we construct a GradleDslLiteral where the
name is not obviously redundant once the value has been deleted.  Most
obviously, this happens in calls to testInstrumentationRunnerArgument,
where the name and value are a pair of arguments to the function.
Deleting the nameElement's PsiElement explicitly resolves this case,
while being harmless if the name's PsiElement has already been deleted.

In order to glue everything together, make sure that the special-case
handling of testInstrumentationRunnerArgument and the implementation of
addToParsedElementMap are consistent in their treatment of model effects
and version constraints.

Bug: 194192720
Test: existing, added
Change-Id: I95175f70152053cbaecdb6460450cdbdb81c1ee4

GitOrigin-RevId: 05223f1f52460135a7982eb2cde79ed7a4b88092