Encapsulate Code Navigation Threading
There are some specific threading constraints that must be respected in
order to use the IntelliJCodeNavigator. One of the two functions handled
the restrictions inside of the code navigator while the other required
the caller to do it.

To make this consistent, this moves all the threading specific code to
be within the code naviagator implementation.

Change-Id: Ifd81649439fcad8c30e36e69d4986653dabbd34e