Group Together All Profiler Preference Variables
There are three variables used to select a process to profile when we
don't have a selected process. This change groups them together in a
preference data type.

This makes the connection between the variables clearer and avoids
needed to have "preferred" before each variable.

By wrapping the values, we can more easily enforce specific
restrictions. The most notable is ensuring we have a non-null filter,
allowing us to skip null checks before using it.

Change-Id: I4e9f6fc2f4c108b7cb558eeaaac781e7dbf3c15f

GitOrigin-RevId: c729b60b00701b84f9eae7ba4f41f99d3d4bb0bc