[BA] Do not show corresponding text when task is not on critical path
At some point we started showing that task determines build duration
for all tasks. This line is always presented on 'Tasks' view as we only
show tasks that determine build duration there.
But now we show this panel also for any task warning and there some
tasks might be not from the critical path thus it is incorrect.

Bug: N/A
Test: added test for the case
Change-Id: I138573cac577ebff814db6492f9607e1337e652d

GitOrigin-RevId: 54169fde0357104c6482fb84b37e2a24cf210795