Fix Compose Preview deploy from the gutter
In the refactor we've made in change Ic2e1bfa8 to update ui-tooling package
references in Android Studio, we accidentally broke the flow of deploying
Compose Previews from the gutter.

The problem is we're not setting the global scope when searching for the
PreviewActivity class, therefore the Compose Preview Run Configuration will
be invalid, since it will only look for this activity within the project
scope, when it's actually located in ui-tooling.

Bug: 196232682
Test: ComposePreviewRunConfiguration*Test should pass + Manual
Change-Id: Ifa15d50084febc6459cdaef3a458f9c79e3294fb