Don't mark as CONNECTED until model update
Previously, AppInspectionInspectorClient considered itself to be
successfully attached (by completing the future returned by doConnect())
as soon as the StartFetchCommand got a successful response. However,
it's still possible that the inspector would fail to ever send any
updates, and if that were to happen we should still time out and fall
back to the legacy client.
In this change we delay setting of the future returned by doConnect()
until a model update happens.

Test: updated existing tests
Bug: none
Change-Id: I575358cdfdadd4764724a15437a8b32443fd3795

GitOrigin-RevId: 8e2cb76833c266a42f1e9f493268c29e05ac97af