Add a Logcat Header Panel
The header panel will have device & client combo boxes similar to the existing Logcat tool.

This is still WIP as there are no UI mocks yet. For now, it uses the DevicePanel code but later, might change to use new components if DevicePanel doesn't meet our needs.

The UI here is temporary and is added just so that we can have a way to select device during development.

Bug: 191880898
Test: Not added because this UI is temporary
Change-Id: I10a83573096168e750be5d90130f01ec96d9edbb