Split Interactive and Animation preview settings checkbox
We were using the same checkbox in the settings panel (Experimental >
Compose) to enable/disable both Interactive and Animation Preview. As part
of recent user studies, we learned that we should make a clear distinction
between these two different tools, and having a separate checkbox to enable
and disable each feature would be more aligned with that goal.

Bug: N/A
Test: Manually verified + searchable_options_test
Change-Id: I79d5171e9473f6b50729f35c461799c8da083f82