Move Code Navigation To Its Own Module
Since code navigation is shared between the inspectors and the
profilers, this change moves it to its own module called
"codenavigation".

While "inspectors-common" was being used as a means of sharing code
between the inspectors and profilers, having a code navigation module
better communicates the scope of the code within it.

Change-Id: I2a622d48e422b986c631085b1478715a8593fff6

GitOrigin-RevId: 9d1726deeb2f7024fddbd14fbcd566a11a0c1c86