Show "Java code" for collapsing java nodes
Java nodes used to be under a `.jar` file under `/apex/*`.
This change makes it more obvious for programmers that they
can collapse Java code.

Code-wise, this change renames methods from saying "path" to
"tag", "frame" to "node", and "hide" to "collapse". It also
computes the tags once and save in the node, instead of saving
the raw path and matching every time.

Bug: n/a
Test: update existing
Change-Id: I980333d5a1bdee9d31c17c5d49ba1791ae9d3a3b

GitOrigin-RevId: a010ab6d19afb5e22c04ad2a1bbb6823c38ebd35