Change JSON Schema For a More Generic Use Case
Instead of having a list of colors and a corresponding list of booleans that tell us if the foreground color should be black or white, we'll have a list of explicit background and foreground colors.

This not only serves the purpose of being more generic but it also better associates the main color with the related boolean.

The getColor() and getFontColor() methods will be renamed to getBackgroundColor() and getForgroundColor() in the next cl.

Bug: 197555112
Test: Updated
Change-Id: Icca06f2cc9d421c975789d1681ce6170a0613b42

GitOrigin-RevId: e094870c17825e9e57bbd086b274209c0ea7b079