[AUA] Initialize recommended version from argument
Rather than unconditionally selecting (initially) the latest known
version (as acquired from the provider) to be upgraded to, allow the
caller of showContent() to specify an initial suggested version.  At the
moment, all callers specify the latest known version anyway, so this
should cause almost no change in behaviour, but this will change
shortly.

"almost" no change in behaviour?  Well, some of the tests used an
inconsistent setup: defining the project to use a particular version of
AGP, and then passing through a different version as the value of a
`current` parameter.  Adjust the tests where this matters.

Bug: 189388326
Test: existing tests adjusted
Change-Id: Ifbb9c7159c7442bf943b34b21566ccb52a9910b9