Make DataVisualizationColors a Class And Rename it to ColorPaletteManager
In order to make it reusable, make DataVisualizationColors a class instead of an object.

Rename it to ColorPaletteManager which describes it function as a general purpose object instead of DataVisualizationColors which describes the current sole use case in the Profiler module.

Since DataVisualizationColors was used from various static locations, it's necessary to provide a static instance. We do this introducing a new singleton DataVisualizationColors in the profiler module.

It was also needed to add ColorPaletteManager as an argument to the static methods in SystemTraceNodeModelHChartColors in order to make it testable.

Bug: 197555112
Test: Updated
Change-Id: I2e6d361ac09535c5cf974ce82639111d2eba2ae6

GitOrigin-RevId: e82254c4b6321d2b5c4cfed0a64d78525386e8c2