[AUA] Revalidate content pane after expanding the tree
We expand nodes and hold the tree size in order to make sure that our
tree pane has enough space to display everything.  However, doing this
at the relatively low level does not inform the parent panes of the
changed space requirements, leading to the wrong size being displayed
until the next repaint.  Force the repaint to happen by revalidating the
content pane.

Bug: none filed
Test: manual
Change-Id: If117fc72b2343b4395ac98c8ca54744af1604382

GitOrigin-RevId: 71d03808020b29c7430b33b2c97179c8ca0b5eda