Add two util methods to ModuleSystemUtil
Module.isHolderModule - a small helper to avoid a == this check when
using Module.getHolderModule

We also move Module.getHolderModule to AndroidModuleSystem in order to
break Gradle dependencies. Other method were not moved in this change
but will be moved once they are used outside of Gradle specific code.

Project.getAndroidModulesForDisplay returns a list of the holder
modules, these should be used where a list of Android Gradle Projects
need to be displayed or the source/dep information is not needed.

Bug: N/A
Test: intergration testing will cover this method once used
Change-Id: Id31f2becd1b891ac29b71a39c46f492717af783c