[AUA] Preserve existing selection even more
We want if possible to avoid removing the user's selected item from the
combo box model, as otherwise adding elements back to an empty model
will trigger one selectedItem update before the selected item is reset
explicitly; then, if we are unlucky with scheduling, the updates to the
model might trigger creation of a processor with a `new` version
different from the intended (previously selected) one.

When updating the model from the suggested versions, therefore, remove
all elements other than one equal to the selected version, if any,
before adding the suggested versions maintaining the sorted state (and
not adding an element equal to the currently selected one, since that
will have been preserved).

Bug: none filed
Test: manual
Change-Id: Ia91954d871ed0e0cd4c0f3e17294907ba4eb547b

GitOrigin-RevId: 890626c351e12a2442b37b29b02941e4700ca09b