Focus the first digit when showing the pairing code dialog.
The previous implementation wrote to a protected field,
myPreferredFocusedComponent; however, this field is only read on Mac:
see DialogWrapper.getPreferredFocusedComponent(). Overriding that method
appears to be the recommended way to set the initially-focused component.

(Focus is ultimately requested by DialogWrapperPeerImpl, which in tests
is replaced by HeadlessDialogWrapperPeer; I tried updating
HeadlessDialogWrapperPeer to similarly request focus, but the focus
request is rejected by Swing.)

Bug: 197988273
Test: manual; focus does not behave the same in tests
Change-Id: I8d4af246570bfe7a1b8af3676f36f0fa3cf9d014

GitOrigin-RevId: 6e8ddce50996ee56c99b909bdecb5957e757d397