Make Right Click Menu Work
The current way of handing right-click in SplittingPanel doesn't work. Once the child component is populated, it "hides" the mouse handler.

Replace it with a system that passes an ActionGroup from the SplittingPanel to the child component. The child component can then do whatever it needs to display it, based on it's implementation details.

In the case of Logcat, the implementation is an EditorEx so we use EditorEX# installPopupHandler() to install a popup handler that contains the actions provided by SplittingPanel.

Bug: 191880898
Test: Added
Change-Id: I3f15c3931474ff227e6ce30b35752d6a0f60fd86

GitOrigin-RevId: 2cf9910aa4b304fa84a6fdd09b74a26fc8a06daf