Set `popupTrigger` correctly for a MOUSE_PRESSED Event
Instead of mouse click.

Components that use com.intellij.ui.PopupHandler actually trigger on either of these but the real UI (on Linux at least) seems to only set it to true on press.

The EditorImpl class only pops up on press:

http://cs/studio-main/tools/idea/platform/platform-impl/src/com/intellij/openapi/editor/impl/EditorImpl.java?l=3879&rcl=e55ce55b980ac887be4c9b693849b93ba75498a6

According to this:
https://docs.oracle.com/javase/tutorial/uiswing/components/menu.html#popup

The exact gesture that should bring up a popup menu varies by look and feel. In Microsoft Windows, the user by convention brings up a popup menu by releasing the right mouse button while the cursor is over a component that is popup-enabled. In the Java look and feel, the customary trigger is either pressing the right mouse button (for a popup that goes away when the button is released) or clicking it (for a popup that stays up).

Fixes: 198157696
Test: N/A
Change-Id: I3b1f5dfbb25ab3d1f4ced765e6fb0e9186a36882

GitOrigin-RevId: df15332e21799064e1f7902288c39fc385b2f410