Fix the Android Studio UI freeze issue
There were bugs reported regarding an UI freeze issue. After an investigation, we identified that a deprecated function - runRenderAction in RenderService could be the cause. The reason why this function was deprecated is that this function could potentially cause a deadlock (that could make the UI freeze).

In this CL, we refactored the codes to adopt a new function (runAsyncAction) that is executed in an asynchronous manner to prevent the deadlock condition. For most of the updated getter functions, we used a cache-like mechanism to store the values so that we don't have to wait for the async function to be completed to obtain the value.

For some of the functions (e.g., getKeyframePos and invoke), we still have to wait until the async function to be completed to ensure the Motion Editor works properly.

Bug: 173103494, 155682989, 151749795, 173526774, 145685089
Test: Manually tested
Change-Id: Idccdbe6263e536b22d3acc4a88c8f62d7f030bd6

GitOrigin-RevId: d4fc8349faf3aef66f6ce64547c2ecb48e1affeb