[Dsl] Correct syntax for proguardFiles
Here we have a property with a varargs adder function for normal use,
but this varargs function does not accept a collection or iterable
argument for setting: thus attempts to setValue() to a reference end up
emitting invalid syntax.

Define a transform to handle this case, and attach it to the
GradlePropertyModels corresponding to proguardFiles and
setProguardFiles.  Setting the value of the property to a reference uses
the setter, while converting it to an empty list changes the name to use
the varargs operator.

Fixes: 72853928
Test: added
Change-Id: I3464f5196bab86c359d4d6fdff833688a0c3730f

GitOrigin-RevId: b6dcb1b0fec80dd80ec7f9468c70399b683ffb46