Refactor logic that create coroutine debugger panel
The current implementation has some timing issues between
our ExecutionListener and the XDebugManagerListener inside
DebuggerConnection.

More in detail, when an app is launched twice, the second time
this happens:

1. ExecutionListener#processStarting is called, which creates
a new DebuggerConnection (at this moment both the old and new
one exist)
2. XDebugManagerListener#processStopped is called, which
calls dispose on both DebuggerConnection.

This results in no panel being showed any app run after
the first.

This change adds a XDebugManagerListener, which is responsible
of creating DebuggerConnection and of forwarding the
processStarted call to DebuggerConnection#processStarted.

Test: NA
Bug: 18202318
Change-Id: Ie729d715113fc081fb05c7df5dfa000fc69d9a10

GitOrigin-RevId: be0a9b8c3d5ca95e2a18c4ba831c69932d3bd6cc