Fix @VisibleForTesting constructor parameter annotations
Placing @VisibleForTesting on a parameter, even if it's a constructor
val parameter, will place the annotation on the parameter itself, not
the field, getter or setter of the corresponding property. For that, use
@field:VisibleForTesting, or @get:VFT or @set:VFT.

This CL updates a handful of places where this was probably the intent;
having @param:VisibleForTesting (which this maps to) is probably not
intentional.

Test: Existing
Bug: None
Change-Id: I8aa008acb51940dfc35189d7fd7f5a6129683421

GitOrigin-RevId: 01a78b18a6bbe0684a97ad842787524fd4efd479