Dispose Properly
The proper way to dispose a child component is to register it with Disposer rather than to call child.dispose() from the parent dispose() method.

The problem with the later is that just calling child.dispose() doesn't alert the Disposer that the object was disposed and things like coroutine cancellation do not work.

Bug: 199391779
Test: updated
Change-Id: Ibd9b64aa52943fcd8e5e8693eecc98e80601cced

GitOrigin-RevId: 69c6684205a133a1af679e622b5a6ee774f20d18