Fix preview blurriness in HiDPI screens
When running on HiDPI screens we removed the slow and expensive scaling
algorithm. However, that caused the image preview to degrade
significantly in HiDPI screens for low levels of zoom.
This change, makes the intermediate preview buffer to be HiDPI aware, so
the image is kept as 2x (on mac) and scaled down from there. That
reduces the blurriness of the image while keeping the speed benefits and
just a bit higher memory consumption.

Test: Manually tested
Fixes: 195035502
Change-Id: Ib166d5361f57142e081ba8cf12963e015e817d34