Fixes issue related to multiple conversions between regular and virtual files
There is a NPE produced by VirtualFile.contentEquals method in DefaultRecipeExecutor.kt:645, as a result of using the LocalFileSystem.getInstance().findFileByIoFile(file) method to look for a virtual file using the a regular file that was converted from a virtual file before using <vFile>.toIoFile(). The issue seems to only affect paths inside of JAR files (jar://<path>).

Bug: 199817138
Test: Manual reproduction of the steps indicated in the issue
Change-Id: I93a223fa8e95576afa6a145f5c6d0d8741939cd6

GitOrigin-RevId: 1036c7b43e84d7d90c13163cb87388c55ce5f558