Refactor LogcatDocumentPrinter
The main incentive for this refactoring is a cl downstream where it became difficult to test the added functionality because LogcatDocumentPrinter was doing too much. Specifically, it was interacting directly with the UI (Editor/Document) and offered little visibility into how it was doing it.

This cl breaks down the functionality of LogcatDocumentPrinter into 3 parts:
1. MessageFormatter formats messages into a .TextAccumulator
2. LogcatMainPanel#appendToDocument() appends a FTextAccumulato to the UI.
3. MessageProcessor: Sends and receives data to/from the coroutine channel.

This not only allows the code downstream to be testable, it results in the existing tests to be a lot easier to perform.

Bug: 191880898
Test: Added/Updated
Change-Id: If27a502a438e12f143c3edea2d9d553801a71a85

GitOrigin-RevId: a2f85fc48e680777e54664ad5af0c244a7801344