Move BuildMap model builder to the IDE codebase
This is to apply it to all Gradle projects (including non-android ones)
and make the model available for projects using older AGP versions.

At this stage this is still a direct replacement of the model builder
provided by the current AGP dev version. In upcoming changes the model
will be requested for root (`:`) projects only and regardless whether
they are Android projects or not.

Bug: n/a
Test: existing (IDE model snapshots)
Change-Id: I0f14fabbd70ff250246e50b2f1f23c8f91f68ddf

GitOrigin-RevId: 3b8a819015c9882d3368f2ffc68a877ee4f48390