Apply Back Pressure on Logcat Reader Process
Since logcat backlog can be really big (over 1 million log entries under extreme cases) we have to throttle the reading thread. The code that consumes the messages is eventually blocked by the EDT which is the bottleneck and if we don't block the reader thread, it may overwhelm the queues feeding the rest of the flow.

I've actually seen this in action. If I had 3 or more logcat windows open all reading from a file with 1.5 million messages simultaneously, Android Studio will very quickly freeze and report memory pressure.

In order to make it possible to read multiple large backlogs of data simultaneously, the reader process cannot be allowed to run unchecked. We must make the data flow block it (when the reader code becomes coroutine friendly, it will suspend rather than block).

When the reader thread is blocked, it will in turn block the output socket of the process which will in turn be blocked by the OS.

There are 2 places we need to block:

1. LogcatReceiver uses a sequential Executor in order to be able to handle the last message in a batch. We still need this executor, but we need to make processNewLines() a blocking call. We do this with a CountDownLatch.
2. LogcatReceiver sends its data to a listener in LogcarReader. This listener is where we switch to coroutine code. We need this coroutine entry point to be blocking so we can truly make LogcatReceiver#processNewLines() blocking.

Bug: 191880898
Test: Since we plan to make this code use coroutines soon, no tests are added at this time.
Change-Id: I962582d198cfdeed1876d8c2b8df21691d0e3495

GitOrigin-RevId: 58e704a47d222d1752313f2b806d9fcd2cf3abfc