[Dsl] Move applyDslElement from GradleDslFile
The ability of Gradle Dsl files to apply other files only applies :-) to
build and settings files, not to arbitrary files which participate in
the construction of the Dsl model.  At the moment the only other
concrete subclass of GradleDslFile represents properties files, which
are not explicitly parsed and in any case are unlikely to have
properties named `apply.<something>`, but we plan in the future to
support Version Catalogs, and to parse properties files explicitly, so
push the handling of apply-as-inclusion into a new abstract superclass
of GradleSettingsFile and GradleBuildFile.

Note that the ability of Settings files to apply other such files is
currently untested.

Bug: n/a
Test: existing
Change-Id: I1d907870e4d0e7535aa026f596fdf361bf18b34c

GitOrigin-RevId: 6b57d81e43e1d61a656b7da4a5e55cf4afd80fd7