Do not cache timestamps lazily
The ProjectSystemClassLoader will keep a list of loaded classes to be
able to do 'is up to date' checks. The timestamp would be added into the
cache lazily (not at the class loading time), so there was cases where
the class had been modified in the meantime and the contents would not
match what it's been seen in the preview.

This change will ensure that the timestamp is obtain at the load time.

Bug: 189559350
Test: Added regression test and checks for isUpToDate
Change-Id: I2a2c8f82e731fdb0cd00ed06addf455323971c95

GitOrigin-RevId: f6311b0040a2fdfcaf3c945c5bd9aaf69a946569