[AUA] Remove buildToolsVersion if it would be ignored
New AGP (by which we mean "since 3.0" :-) will ignore a too-low
specification of buildToolsVersion, replacing it with its minimum
supported version with a warning.  We don't have a convenient table of
data to support removing buildToolsVersion in all cases, but for
upgrades to AGP 7.0 and 7.1 we can remove buildToolsVersion if it will
be below the minimum version supported by those AGPs.

Fixes: 183650762
Test: added
Change-Id: Idb34b09174e7344262cd956d75ff909607f00073