Trim text in PluginDescriptorTest
In IJ 212, the text returned from IdeaPluginDescriptor.getDescription
is no longer trimmed. This is due to IJ commit e63d824c05, which
removed the call to 'getChildTextTrim' when loading the description.

This appears to have no effect in production, since descriptions are
already wrapped in HTML before being displayed to the user. We just
need to update our test.

Change-Id: Icfe38e2a9b07159ff5a010ae7af4daba4ff52e6f

GitOrigin-RevId: 5364dfd7ddd672ac4f5cdd0ad277867123ba952b