Replace Explicit Width/Height With Preset Size %
Having to specify width and height explicitly is not a reasonable API, especially since there's no expectation the user even knows the dimensions of the device without having to check.

I tested various %'s and found that 75% actually creates a larger file, 50% results in about 65% file reduction and 30% in 25% reduction. So I;m setting the presets at 100, 50 & 30

Fixes: 154394433
Test: added
Change-Id: Ibc7626369c4389986c4e0d792d0301c517a7d9fc

GitOrigin-RevId: 80486bf5eb1ee09ce37d95a0fcaf0a0cc381250c