Remove cleanTest task from androidGradleTesttasksProvider
Intellij made some Platform changes to always force test re-execution,
that makes the usage of cleanTest task no longer needed.
The way they do this is by enforcing Task.UpToDateWhen to be always
false in the Gradle init scripts.

This changes deletes the usage of cleanTest task. We also add a test to
verify that the test will be re-executed even if there was no sources
change.

Bug: 194989584
Test: Added
Change-Id: I83b3aeeceb3b222ab9525868076bb605ff4bf8e0

GitOrigin-RevId: 40416e8251e1dc45c1ac7519ed7775d00d2dea46