[Dsl,.toml] Simplify visitor
Well, simplicity is in the eye of the beholder.  But if we have the
visitor maintain not only Dsl context but also the name of any key we
have just seen, we can arrange for each visit method to do essentially
one thing: the Literal / InlineTable / Array visitors construct and add
a parsed element to their context, named by whatever name has been
seen in a surrounding KeyValue.

Bug: 200280395
Test: existing
Change-Id: I0d3a4009d2e2cde04f94e655ac7666f861158f55

GitOrigin-RevId: 0ab84d6fcc467d7072d7c3527fd07921ed342146