[Dsl,.toml] Support version.ref in Version Catalogs
In generic TOML, there's no such thing as a reference.  In Version
Catalogs, library / alias versions can indirect to the `versions` map
by defining version to be a table containing a `ref` key whose value is
the name of the key in the `versions` map; likewise for plugins.

We should not do anything special at the Toml parser level to handle
these references -- though if requirements change such that we have
to, we should only do so if the dslFile is a GradleVersionCatalogFile.
For the moment, we fix up version references after the fact, replacing
the map expression with a suitably crafted literal, and ensuring that
the injection and dependent states for that literal and its target are
correct.

Bug: 200280395
Test: needed
Change-Id: Ie8b08b7d60c3892a3d39b33f06ea561227c78f78

GitOrigin-RevId: e9b74e99521d6f2a77c1838e40d6830ce23c3d66