[MPSS] Add check for holder module in findAndroidFacetForTests
AndroidGradleTests.findAndroidFacetForTests looks for AndroidFacet
to set up in test. It is possible to request by module name,
this makes sense to look over all modules.
Otherwise it tries to find non-lib module and in this case it makes
sense to look only at holder modules because any module of the group
can match here whichever happen to be first in the list.

Bug: N/A
Test: N/A
Change-Id: Ifff698f4559c49be2103cc38efdbf2aee745b4e2

GitOrigin-RevId: 903011c8daeef4b827531b544669e9f515317088