[MPSS] Fix android run configurations
Existence of additional modules for source sets should be transparent
for the user. We should show on the ui (including module selectors)
the holder modules and resolve them to specific source set modules
on the background where necessary.

From here comes the decision to keep holder modules in run configuration.
First of all this means that we should filter holder modules in module
selector in AndroidRunConfigurationEditor.
This also brings the benefit tht no migration is needed, all previously
saved configurations will work as before as nothing would change in
their state.

The only challenge here is adjusting AndroidTestRunConfiguration properly
as it does various checks and suggest aid for the class selector. This
happens on the various levels in the components used and require careful
review to amend search scopes and used modules accordingly.

AndroidRunConfiguration and AndroidTestRunConfiguration share the same
AndroidRunConfigurationEditor.
Previous fix filtered only main modules in the selector but this does
not work for android tests. Android tests requires .androidTest module
that contains the target classes to be selected otherwise
validation does not pass.

This change makes it possible to pass modules filter from the outside
and thus choose which modules to filter on editor creation.
AndroidRunConfigurationModule redefines search scope for test
configurations to process find class requests properly.
Additionaly, AndroidRunConfigurationModule overrides setModule method
to ensure holder module is set to configuration even if source set
module is coming from the action context.

This change was manually verified that all basic functionality
of android run configuration (creating from inside the dialog, creating
from the editor and the project view, selecting test package/class/method
and corresponding checks in the editor) keeps working on both modes
and the transition between the two modes is smooth.

Bug: N/A
Test: Manually verified, sync snapshots updated.
Change-Id: I4415e969ca5f420969cff802d9ebc2064f0adf6c

GitOrigin-RevId: 8f845e93e25147c840d73bd437fce79f758b1ff8