Use models, not components, for issue description
Using components to remember which preview is affected by an issue is
not reliable for Visual linting. For example, when using RecyclerView,
we don't have access to the components of each item. Also, if the issue
is happening for each item, there could be an overcount.
So in addition to remembering affected components when available (for
highlighting them), also remember the models that are affected, and use
the number of affected models in the issue description.

Bug: 201571675
Test: checked on trackr
Change-Id: Ic47713a40a9d05204b13f677269358819d7714fb

GitOrigin-RevId: d9340d9aef74bf5bbbeed7b411995aad867f30e4