Add support for various formatting options
1. Enable/Disable fields (timestamp, PID/TID, tag, appName)
2. Timestamp as date-time or just time
3. PID only or with TID
4. Tag & app name max length
5. Tag & app name hide duplicates.

This replaces the current behavior of adaptive width for tags and app names. This strategy doesn't work well in the real world where an outlier tag or app name can set a very large length, and we get stuck with it.

A dialog that configures the log will be added downstream.

Bug: 191880898
Fixes: 201666978
Test: Added
Change-Id: I7d8d0bf673f5d33088998d158e52a534f2e28471

GitOrigin-RevId: 2de7013e486bffe632f1780a90d8355177a4d2f1