Add a Message Backlog
The purpose of this backlog is to allow for filtering and re-rendering of the Logcat when needed, for example, when formatting changes are made.

The message backlog size made to be smaller than the actual space it takes to render it in the editor so there are at least as many messages there as there are in an unfiltered log window.

Bug: 191880898
Test: Added
Change-Id: I544c3e2d05c94896fdc3ca5c4be3421e529e905f

GitOrigin-RevId: a4a61ee945a753a4fccb32fe70c0877493b1cbad