[MPSS] Make GradleSyncIntegrationTest work with source sets
We change the modules used in these tests to work with the main modules
for both Android and Java Gradle projects. We add a new test utility to
aid us in getting the main Java modules when mpss is turned on.

We also remove an assertion on modules size from one of the tests as
this is covered by the snapshot tests and we don't need to reassert it
here.

Bug: N/A
Test: N/A
Change-Id: I45e21873ebf5b4b3f4a055cfc1384dd15edcba28

GitOrigin-RevId: 00c24c6fa1def380578c047a2a137c485d363845