[AUA] Upgrade single-interpolation version declarations
For the case of classpath dependencies a version that is a string
interpolation of a single variable as in "foo:foo:$fooVersion" is
managed by the ArtifactDependency model; for the case of plugin
dependencies in build files, the form `id 'foo' version "$fooVersion" is
illegal.  However, that latter form is legal in pluginManagement plugins
blocks, and it would be easy to write rather than a bare variable
reference, so support it in the upgrader.

I don't think it will ever be reasonable to attempt to support more
exotic forms, like "${fooMajor}.${fooMinor}"; we would have to implement
unification, which is quite a lot of work for minimal gain.

Bug: 201946668
Test: added
Change-Id: I5fefc7f3c72a6651366541ce7b297903d7bf336a

GitOrigin-RevId: 868f3df85d7de3021c27ed25df0e4bed3f43e661