Remove Micro Optimization Prebuilt Strings
The transient fields we not being initialized properly by Gson.

This was causing misalignment when the tag/app length was different from the default because Gson would initialize the transient fields from the default constructor and was not updating them when it initialized the other fields.

The optimization is small anyway and not worth going through the trouble of solving this, for example, with a @JsonAdaptor.

Bug: 191880898
Test: Removed
Change-Id: Idfc8578dfd3f21a176a68e2380153af858075f69

GitOrigin-RevId: 2e5d8abe593b8b49553bf8a51f4dc2cf5cebefec