[Dsl,.toml] Accessor normalization for Version Catalog aliases
Because version catalogs allow aliases which do not correspond to
identifiers, but the expected way to use those aliases is through
type-safe language accessors, we need to mimic the normalization that
those aliases go through, treating `-` and `_` characters as hierarchy
separators.  The format of the version catalog already requires that
each part separated by those characters has identifier nature, so we do
not need to check for strange characters within each split.

Emulating the name mangling requires a new interface method,
setNameElement(), which is only useful for the purpose of updating a
previously-assigned name in the light of name mangling or other
context-dependent semantics.  Document the fact that this is very
unlikely to be the method that a client needs.

Bug: 200280395
Test: added
Change-Id: I21ffc925ff80de4686c735cff63c39e8e32c1cee

GitOrigin-RevId: 7539fe0420bc179143cc09e935440b490d15b1b1