[AUA] Call the right entry point from tests
The entry point is the three-argument form, which delegates to others.
This didn't matter until the built-in rules implementation changed in
version 2021.2 of the platform, which meant they were no longer returned
from one of the delgated methods.

Also note that the new implementation no longer has all the built-in
groupings with rank of Integer.MAX_VALUE; our defensive choice of
getRank() is still fine.

Fixes: 202330008
Test: existing tests adjusted
Change-Id: Ibd26cca15eaf133871d82fc75b838480ad339b6f

GitOrigin-RevId: e08d380a686224570d13db1a7e8165df863ca261