[AUA] Move the okButton closest to the combo
This may not be an entirely compliant ordering for normal components,
where I think there's a platform dependency on whether the default
action goes to the left or the right.  But this UI is probably
sufficiently off the usual path -- we're not in a dialog box, Toto --
that the need for clarity supersedes platform norms.

Bug: 200975232
Test: manual
Change-Id: I51703a603cd1a5de007bdf6e4290304ecdc58fa5

GitOrigin-RevId: 89fcae6fbc36ee3b8f617a92334fc6f76102a334