Show a warning when Previews with ViewModel fail to render
We recently added support to render Compose Previews tha receive a basic
ViewModel, i.e. a ViewModel that does not trigger operations that are not
supported in layoutlib, e.g. network access and database connection.

Although we have support for such ViewModels, it's still not a good practice
to use them in Compose Previews, as ideally the logic code (ViewModel)
should be decoupled from UI code (Composable). When we fail to render a
Compose Preview that receives a ViewModel, we now show a warning that
ViewModels often trigger unsupported operations and link developers to the
Compose Tooling DAC page, where they can learn more about Compose Preview
good practices.

Bug: 199834697
Test: ComposeRenderErrorContributorTest
Change-Id: I15a1584106ba4fe45b38cd2c32d0831c21f01129

GitOrigin-RevId: f3b057414c79fa5e5932a8a377c2e3a243e43dce