[Dsl] Implement a transform for plugin names
Rather than finding the name element specially when creating a
PluginModel, defer that work to when the name model is actually used by
adding a specialised transform.  This will simplify the integration of
version catalogs, when it becomes allowed to use references for the
names and other attributes of plugins.

Bug: 200280395
Test: existing
Change-Id: Idb1a1542da86c9a5a8bce3e452bd5624faf0ff21

GitOrigin-RevId: a0576ffa8dfae75dc06ee835f98397b434a22b0e