Fork StackTraceExpander & ExceptionFolding
When Logcat V1 goes away, we will still need these files.

Convert them to Kotlin refactor to handle the entire batch of lines instead of one line at a time. We can do this because we have access to an entire LogCatMessage that contains all the lines in the stack trace.

This makes us not need to keep a state, so we can convert the class into an object.

We also need to disable the Logcat V1 ExceptionFolding extension if not applied to a Logcat ConsoleView, so we don't double process.

Bug: 191880898
Test: Added
Change-Id: I0108a8cc1fa9693a1b867062bcc5dd47c180a7f6

GitOrigin-RevId: 61e418fb2eb5ce1182db36018eff6b58cddfb8c8