[MPSS] Fix findAndAddGradleBuildTasks for mpss
For Java module under mpss the dependency we follow from Android modules
is the main source set module. This module does not have either the
GradleFacet or JavaFacet on. As such we need to check the parent module
to see if we can find a facet there or not. This will be needed until we
remove both the Gradle and Java facets and pull the required information
from intellijs platform apis.

We also ensure that GradleModuleTasksProviderTest uses the correct
module.

Bug: N/A
Test: N/A
Change-Id: I3a48a7c643bd0cdd3a739ebd26c8ede644a43774

GitOrigin-RevId: 639bff7b3bea5ea2ea4be644d03cdd779d95db95