[BA] Jetifier UX improvements
Change the header text to be more self-explanatory and include a call
to action on each step.
Present checkJetifier result on a two-sided table-tree component:
Left table part lists declared dependencies in the project that
detected to be depending on the legacy support libraries.
On selection right tree part shows the path from that support
library to this declared dependency.
Make the jetifier warning page layout be the same for all the
steps just changing the content corresponding to current step.

Bug: 195915300
Test: Added
Change-Id: Iba783dac9c1f7ab012987db775eb7edf288a1ec2

GitOrigin-RevId: 68df114f9986f6636c2f12629e162137bd5f50ea