Avoid RejectedExecutionException in tests.
We used to create a Executors.newSingleThreadExecutor to pass to
LayoutInspectorRule. The problem happened because the execution on
the thread would sometimes happen after the test was done and the
the thread pool was closed.

Replace the thread pool with the mechanism used in
AppInspectionInspectorMetricsTest and move this logic to
LayoutInspectorRule.

Fixes: 203083847
Bug: 202751425
Test: DeviceViewPanelTest
Change-Id: I1361ccddb86ad1db580a922bc73d2d1b3f892138

GitOrigin-RevId: 3e741ec4640208aaf7a47d807b72d77e9bd0cf3e