[Dsl] Improve Kotlin project configurator for Android
The Kotlin project configurator, in general, cannot work on a single
file at a time: given that there are multiple ways of specifying the
Kotlin plugin's version, in various different places; multiple places
where repositories can be added to a project's dependency
resolution (and build failures if they're added in the wrong place),
the Kotlin project configurator must operate with full knowledge of
the entire build for the project, not just a single module's build
file.

Therefore, rewrite the configurator to use the Gradle Build Model for
its build file manipulation.  This allows us to do the right thing for
our current templates as well as older ones, placing repository
declarations where necessary.

A number of unit tests of the Android Gradle Kotlin configurator now
have changed output, almost entirely cosmetic.  (The test itself is a
little bit misguided, in that it operates on a single build file
without a wider project context, which is necessary to be certain of
correctness.)

Bug: 180411529
Test: existing tests adjusted; uitest fix in next CL
Change-Id: I7a7500a44db6567664b30ef72d8c3510f1ec2fc8

GitOrigin-RevId: 0c95d3e0d2f76e10603f6336f3573fa7dd4ced76