Do Not Implement DataProvider
For some reason, if LogcatMainPanel implements DataProvider and supports a CommonDataKeys.EDITOR dataId, some basic things don't properly.

Specifically, adding JTextField as child of LogcatMainPanel will not handle some keys (Backspace, Arrows, Enter etc.). Instead, they seem to be diverted to the Editor instead.

The DataProvider was needed in order to support ToggleUseSoftWrapsToolbarAction but this can be circumvented by extending the class to provide an Editor directly.

Bug: 191880898
Test: N/A
Change-Id: I3efb1a0d2e6e3700e1bdf8312ad5adf108ac1fa3

GitOrigin-RevId: 0112be23ec8f146665018da22644365d7c7e64e9