Wait for snapshot before saving
When we save a snapshot in non-live mode, we just save away the data we
received previously to display in the inspector. This data is only saved
when in non-live mode, since in live mode we don't get all the
parameters data each time. However, if we've just switched from live
mode to non-live mode we may not have received the non-live data yet,
leading to NPEs or empty snapshots.
Now we show a progress dialog and wait until the snapshot info has been
received before trying to save it.

Test: Added a test
Fixes: 203206695
Fixes: 191429263
Change-Id: Ic996ec007521e3df55bcd775b3cf6242ad201fc1

GitOrigin-RevId: 2ca2ce8a6739c6a22ac33f8fcdd2877a0c7712d6