Revert "Update model snapshot for testFixtures project"
This reverts commit 809ce7fc3eab7c9675f69d8f1f6bd56e156d4c45.

Reason for revert: variantAttr is no longer used in v2 model, however
in v1, we still need to select the correct variant in case of
testFixtures module dependency. And since testFixtures is not
supported in v1 models, it's better to set the variantAttr to the
main variant instead of the testFixtures variant so the IDE could
pick up the main variant directly.

Change-Id: I5ffa556909c3918e7428243368c7f49e742144be

GitOrigin-RevId: 474e12dee7b1a6535238d35e2159fd06efd1a8f4