Make the animation graph an area chart
The animation graph in the Timeline Panel of the Motion Editor is represented with a line chart. However, the graph is displayed using an area chart in Compose. To make the presentation consistent, we want to update the line chart in the Timeline Panel into an area chart.

To make it look more consistent, we need to draw different parts in the following order - 1. Grid Line, 2. Horizontal baseline where y = 0 and an extra y = 1 baseline for the scaleX and scaleY attributes, 3. waveOffset dashed line, and 4. the actual graph. We will also have 1:1 color mapping for each attribute based with the suggested colors.

A feature flag is also added to allow us to easily switch between the original design and the new design.

Bug: 203778933
Test: Manually tested
Change-Id: I19c189e96c30c905a7db9417d63f2ac0616c18c7

GitOrigin-RevId: c517887b14747b46eb0b6e696980ee3ca16dd78a