Disable coroutine debugger on api lower than 29
On api 27 the agent .so is not found at startup time.
This is probably because there is no guarantee that
the code_cache folder (where we put the .so) is created
during app install.

On api 28 the whole debugger hangs when attaching the agent.
For now we're just disabling, could investigate why later on.

Test: CoroutinesDebuggerLaunchTaskContributorTest
Bug: 182023182
Change-Id: Ia0cff1463a6e3a8601f3bb1bd73824d520620e6f

GitOrigin-RevId: eac1a427d6b8c108821b0adb231fc5cfb8eeaee5