Fix hover color in properties table
Set the hover background color for cells that are selected and the
table is not focused.

JBTable.prepareRenderer assumes a cell renderer will set the color
to the selected table color if the cell is selected.

We (the properties table) are only setting the selected color if
the table also has focus. This creates a situation where the
hover color looks wrong.

Fixes: 203443166
Test: Unit test added
Change-Id: Ifbf682918e53be21dcc35f259df7c24cb1c54f93

GitOrigin-RevId: 3a99e33ae07ad134070dd1e7d1164658f60d100a